Move add-on show commands to display helper
